% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_blank.R
\name{question_blank}
\alias{question_blank}
\title{Fill in the blank question for learnr tutorials}
\usage{
question_blank(
  text,
  ...,
  type = "blank",
  correct = "Correct!",
  incorrect = "Incorrect",
  try_again = incorrect,
  allow_retry = FALSE,
  random_answer_order = FALSE,
  placeholder = "Enter answer here...",
  trim = TRUE,
  rows = NULL,
  cols = NULL,
  style = "tutorial_question",
  options = list()
)
}
\arguments{
\item{text}{Question or option text}

\item{...}{parameters passed onto learnr answer.}

\item{type}{Type of quiz question. Typically this can be automatically
determined based on the provided answers. Pass \code{"radio"} to indicate that
even though multiple correct answers are specified that inputs which
include only one correct answer are still correct. Pass \code{"checkbox"} to
force the use of checkboxes (as opposed to radio buttons) even though only
once correct answer was provided.}

\item{correct}{For \code{question}, text to print for a correct answer (defaults
to "Correct!"). For \code{answer}, a boolean indicating whether this answer is
correct.}

\item{incorrect}{Text to print for an incorrect answer (defaults to
"Incorrect") when \code{allow_retry} is \code{FALSE}.}

\item{try_again}{Text to print for an incorrect answer (defaults to
"Incorrect") when \code{allow_retry} is \code{TRUE}.}

\item{allow_retry}{Allow retry for incorrect answers. Defaults to \code{FALSE}.}

\item{random_answer_order}{Display answers in a random order.}

\item{placeholder}{Sample text to appear in blank.}

\item{trim}{Logical to determine if whitespace before and after the answer
should be removed.  Defaults to \code{TRUE}.}

\item{rows, cols}{Defines the size of the text input area in terms of the
number of rows or character columns visible to the user. If either \code{rows}
or \code{cols} are provided, the quiz input will use \code{\link[shiny:textAreaInput]{shiny::textAreaInput()}}
for the text input, otherwise the default input element is a single-line
\code{\link[shiny:textInput]{shiny::textInput()}}.}

\item{style}{can change display of question to "notes" or "exam"}

\item{options}{Extra options to be stored in the question object. This is
useful when using custom question types. See \code{\link[sortable:question_rank]{sortable::question_rank()}}
for an example question implementation that uses the \code{options} parameter.}
}
\value{
A custom \code{learnr} question, with \code{type = blank}.
}
\description{
Add interactive fill in the blank tasks to your \code{learnr} tutorials.
You must include at least one single blank line ___ in the question text.
A blank is created by typing 3 underscores.
}
\details{
The text utilizes HTML for further display customization, which can be used
to insert an image or line break.
}
\examples{
question_blank(
  "3 + ___ = 5 <br/>
   ___ - 4 = 4",
  learnr::answer("2", correct = TRUE),
  learnr::answer("8", correct = TRUE),
  allow_retry = TRUE
)

}
